+++ modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/internal/content/processor/DefaultTextExportImportContentProcessor.java
@@ -724,6 +724,33 @@ public class DefaultTextExportImportContentProcessor
 
 				urlSB.append(_DATA_HANDLER_GROUP_FRIENDLY_URL);
 
+				// Append the UUID. This information will be used during the
+				// import process when looking up the proper group for the link.
+
+				urlSB.append(StringPool.AT);
+
+				if (urlGroup.isStaged()) {
+					Group liveGroup = urlGroup.getLiveGroup();
+
+					urlSB.append(liveGroup.getUuid());
+				}
+				else if (urlGroup.isStagedRemotely()) {
+					String remoteGroupUuid = urlGroup.getTypeSettingsProperty(
+						"remoteGroupUUID");
+
+					if (Validator.isNotNull(remoteGroupUuid)) {
+						urlSB.append(remoteGroupUuid);
+					}
+				}
+				else if (group.getGroupId() == urlGroup.getGroupId()) {
+					urlSB.append(urlGroup.getFriendlyURL());
+				}
+				else {
+					urlSB.append(urlGroup.getUuid());
+				}
+
+				urlSB.append(StringPool.AT);
+
 				String siteAdminURL =
 					GroupConstants.CONTROL_PANEL_FRIENDLY_URL +
 						PropsValues.CONTROL_PANEL_LAYOUT_FRIENDLY_URL;
@@ -1028,8 +1055,49 @@ public class DefaultTextExportImportContentProcessor
 			content, _DATA_HANDLER_COMPANY_SECURE_URL, companySecurePortalURL);
 		content = StringUtil.replace(
 			content, _DATA_HANDLER_COMPANY_URL, companyPortalURL);
-		content = StringUtil.replace(
-			content, _DATA_HANDLER_GROUP_FRIENDLY_URL, group.getFriendlyURL());
+
+		// Group friendly URLs
+
+		while (true) {
+			int groupFriendlyUrlPos = content.indexOf(
+				_DATA_HANDLER_GROUP_FRIENDLY_URL);
+
+			if (groupFriendlyUrlPos == -1) {
+				break;
+			}
+
+			int groupUuidPos =
+				groupFriendlyUrlPos + _DATA_HANDLER_GROUP_FRIENDLY_URL.length();
+
+			String groupUuid = content.substring(
+				groupUuidPos + 1,
+				content.indexOf(StringPool.AT, groupUuidPos + 1));
+
+			Group groupFriendlyUrlGroup =
+				_groupLocalService.fetchGroupByUuidAndCompanyId(
+					groupUuid, portletDataContext.getCompanyId());
+
+			if ((groupFriendlyUrlGroup == null) ||
+				groupUuid.startsWith(StringPool.SLASH)) {
+
+				content = StringUtil.replaceFirst(
+					content, _DATA_HANDLER_GROUP_FRIENDLY_URL,
+					group.getFriendlyURL(), groupFriendlyUrlPos);
+				content = StringUtil.replaceFirst(
+					content, StringPool.AT + groupUuid + StringPool.AT,
+					StringPool.BLANK, content.indexOf(group.getFriendlyURL()));
+
+				continue;
+			}
+
+			content = StringUtil.replaceFirst(
+				content, _DATA_HANDLER_GROUP_FRIENDLY_URL, StringPool.BLANK,
+				groupFriendlyUrlPos);
+			content = StringUtil.replaceFirst(
+				content, StringPool.AT + groupUuid + StringPool.AT,
+				groupFriendlyUrlGroup.getFriendlyURL(), groupFriendlyUrlPos);
+		}
+
 		content = StringUtil.replace(
 			content, _DATA_HANDLER_PATH_CONTEXT, _portal.getPathContext());
 		content = StringUtil.replace(