+++ sql/portal-tables.sql
@@ -366,16 +366,6 @@ create table DLFileEntryTypes_DLFolders (
 	primary key (fileEntryTypeId, folderId)
 );
 
-create table DLFileRank (
-	fileRankId LONG not null primary key,
-	groupId LONG,
-	companyId LONG,
-	userId LONG,
-	createDate DATE null,
-	fileEntryId LONG,
-	active_ BOOLEAN
-);
-
 create table DLFileShortcut (
 	uuid_ VARCHAR(75) null,
 	fileShortcutId LONG not null primary key,
@@ -452,15 +442,6 @@ create table DLFolder (
 	statusByUserId LONG,
 	statusByUserName VARCHAR(75) null,
 	statusDate DATE null
-);
-
-create table DLSyncEvent (
-	syncEventId LONG not null primary key,
-	companyId LONG,
-	modifiedTime LONG,
-	event VARCHAR(75) null,
-	type_ VARCHAR(75) null,
-	typePK LONG
 );
 
 create table EmailAddress (